# Replace remaining regexps