--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,6 @@ module github.com/prometheus/otlptranslator
 go 1.23.0
 
 require (
-	github.com/grafana/regexp v0.0.0-20240518133315-a468a5bfb3bc
 	github.com/stretchr/testify v1.10.0
 	go.opentelemetry.io/collector/pdata v1.28.1
 )