@ArthurSens 2025-03-27
Oh my gosh I just realized I benchmarked the wrong function 🤦. I'll get back to this later