@ArthurSens 2025-03-27
Ok, description updated with the correct benchmark results